backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
